+++ Source/Synth/Voice.h
@@ -87,6 +87,8 @@ private:
     std::atomic<float>* subLevel = nullptr;
     std::atomic<float>* noiseLevel = nullptr;
     std::atomic<float>* oscWarp = nullptr;
+    std::atomic<float>* oscWavetablePosition = nullptr;
+    std::atomic<float>* osc2WavetablePosition = nullptr;
     std::atomic<float>* ampAttack = nullptr;
     std::atomic<float>* ampDecay = nullptr;
     std::atomic<float>* ampSustain = nullptr;